fix merge

+++ common/ASC.Data.Storage/DataOperators/ReadOperators/TarReadOperator.cs
@@ -27,7 +27,7 @@
 namespace ASC.Data.Storage.DataOperators;
 public class TarReadOperator: BaseReadOperator
 {
-    public TarReadOperator(string targetFile)
+    public TarReadOperator(string targetFile, bool removeTarget = true)
     {
         _tmpdir = Path.Combine(Path.GetDirectoryName(targetFile), Path.GetFileNameWithoutExtension(targetFile).Replace('>', '_').Replace(':', '_').Replace('?', '_'));
 
@@ -37,6 +37,9 @@ public class TarReadOperator: BaseReadOperator
             tarOutputStream.ExtractContents(_tmpdir);
         }
 
-        File.Delete(targetFile);
+        if (removeTarget)
+        {
+            File.Delete(targetFile);
+        }
     }
 }
